Added log line. Made MaxAttestations an u64

+++ pallets/proof_of_existence/src/mock.rs
@@ -31,7 +31,7 @@ impl pallet_timestamp::Config for Test {
 // Poe
 pub const MILLISECS_PER_PROOF_ROOT_PUBLISHING: u64 = 6000;
 pub const MIN_PROOFS_FOR_ROOT_PUBLISHING: u32 = 2;
-pub const MAX_STORAGE_ATTESTATIONS: u32 = 10;
+pub const MAX_STORAGE_ATTESTATIONS: u64 = 10;
 
 parameter_types! {
     pub MaxAttestations: MaxStorageAttestations = MaxStorageAttestations(MAX_STORAGE_ATTESTATIONS);
